% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{BatchXmlText}
\alias{BatchXmlText}
\title{Retrieve xml context}
\usage{
BatchXmlText(xmlObj, xPrefix, xItems)
}
\arguments{
\item{xmlObj}{xml object.}

\item{xPrefix}{xPath prefix.}

\item{xItems}{Node names.}
}
\value{
A named list.
}
\description{
Batch retrieve xml context
}
\details{
Retrieve xml context by xPath.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

