% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epost.R
\name{EPostNCBI}
\alias{EPostNCBI}
\title{NCBI EPost API}
\usage{
EPostNCBI(postPara, ...)
}
\arguments{
\item{postPara}{A named list of HTTP POST terms. For example, "db" is E-utility Database Name, see \url{http://www.ncbi.nlm.nih.gov/books/NBK25497/table/chapter2.T._entrez_unique_identifiers_ui/?report=objectonly}.}

\item{...}{Parameters inherited from the "postForm()" function in the "RCurl" package.}
}
\value{
A names list containing "QueryKey", "Count", "WebEnv".
}
\description{
NCBI Database API - Directly use NCBI EPost API
}
\details{
NCBI EPost provide thousands of queries in one HTTP POST.
}
\examples{
genePara <- list(db = "gene", id="948242,15486644")
genePost <- EPostNCBI(genePara)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
Entrez Programming Utilities Help \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/}
}

