% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurl.R
\name{EUrl}
\alias{EUrl}
\title{E-utilities url}
\usage{
EUrl(eutils)
}
\arguments{
\item{eutils}{The nine E-utilities tools in lower cases.}
}
\value{
url
}
\description{
Generate E-utilities url
}
\details{
A wrap function to generate E-utilities url. For now, it supports "einfo", "esearch", "epost", "esummary", "efetch", "elink", "egquery", "espell" , and "ecitmatch".
}
\examples{
efetchUrl <- EUrl("efetch")
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

