% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid2mRNAid.R
\name{ExtractmRNAid}
\alias{ExtractmRNAid}
\title{Get mRNA accession number from gene id for one gene}
\usage{
ExtractmRNAid(geneXml)
}
\arguments{
\item{geneXml}{Gene xml data.}
}
\value{
NA or mRNA accession number
}
\description{
NCBI Database API - Get single NCBI gene information
}
\details{
Get mRNA accession number.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

