% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid2mRNAid.R
\name{geneid2mRNAid}
\alias{geneid2mRNAid}
\title{Get mRNA accession number}
\usage{
geneid2mRNAid(NCBIGeneIDs, n = 1, maxEach = 10000)
}
\arguments{
\item{NCBIGeneIDs}{A vector of NCBI gene or protein IDs.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{maxEach}{The maximum retrieve number in each visit. The ESearch, EFetch, and ESummary, the max number in one query is 10,000.}
}
\value{
A list containing retrieve mRNA accession number
}
\description{
NCBI Database API - Get NCBI mRNA accession number from given NCBI gene IDs
}
\details{
Get NCBI mRNA accession number
}
\examples{
geneid2mRNAid(c('100286922', '948242', '15486644'), n = 2)

}
\references{
Entrez Programming Utilities Help \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
##' @importFrom RCurl postForm
}
