% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{getNCBIGenesInfo}
\alias{getNCBIGenesInfo}
\title{Get NCBI genes information}
\usage{
getNCBIGenesInfo(NCBIGeneIDs, n = 1, maxEach = 10000)
}
\arguments{
\item{NCBIGeneIDs}{A vector of NCBI gene IDs.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{maxEach}{The maximum retrieve number in each visit. The ESearch, EFetch, and ESummary, the max number in one query is 10,000.}
}
\value{
A list containing gene information for each ID. A empty character vector (whose length is 0) will be returned for the items if the contents are not found.
}
\description{
NCBI Database API - Get NCBI gene information from given NCBI gene IDs
}
\details{
Get NCBI gene information, including gene name, description, genetic source, aliases, gene location. To retrieve thousands of proteins, use EPost to post record into the web server and then retrieve data using ESummary. If the gene ID is not found, return an error information in the list.
}
\examples{
gene3 <- getNCBIGenesInfo(c('100286922', '948242', '15486644'), n = 2)
## not found
ghostInfo <- getNCBIGenesInfo('111111111', n = 1)
\dontrun{
require(KEGGAPI)
## signle genome with two plasmids
smuGenes <- convKEGG('smu', 'ncbi-geneid')
smuGeneNames <- sapply(strsplit(smuGenes[, 1], split = ':', fixed = TRUE), '[[', 2)
smuInfo <- getNCBIGenesInfo(smuGeneNames, n = 4)

## two genomes with two plasmids
draGenes <- convKEGG('dra', 'ncbi-geneid')
draGeneNames <- sapply(strsplit(draGenes[, 1], split = ':', fixed = TRUE), '[[', 2)
draInfo <- getNCBIGenesInfo(draGeneNames, n = 4)
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
Entrez Programming Utilities Help \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/}
}

