% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{getNCBITaxo}
\alias{getNCBITaxo}
\title{Get NCBI taxonomy information}
\usage{
getNCBITaxo(NCBITaxoIDs, n = 4)
}
\arguments{
\item{NCBITaxoIDs}{A vector of NCBI taxonomy IDs.}

\item{n}{The number of CPUs or processors, and the default value is 4.}
}
\value{
A list containing taxonomy information for each ID.
}
\description{
NCBI Database API - Get NCBI taxonomy information from given NCBI taxonomy IDs
}
\details{
Get NCBI taxonomy information.
}
\examples{
tax3 <- getNCBITaxo(c('9606', '511145', '797302'), n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
Entrez Programming Utilities Help \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/}
}

