% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNCBI.R
\name{searchNCBILocus}
\alias{searchNCBILocus}
\title{Search NCBI locus ID}
\usage{
searchNCBILocus(NCBILocusID, type = "gene")
}
\arguments{
\item{NCBILocusID}{A \code{string} of NCBI locus ID.}

\item{type}{A \code{string} of database.}
}
\value{
A \code{string} of NCBI gene ID.
}
\description{
NCBI Database API - Search single NCBI locus ID
}
\details{
Search single NCBI locus ID and retrieve its.
}
\examples{
searchNCBILocus('G4B84_011174')

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
