% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{singleGeneInfo}
\alias{singleGeneInfo}
\title{Get single NCBI gene information}
\usage{
singleGeneInfo(geneXml)
}
\arguments{
\item{geneXml}{Gene xml data.}
}
\value{
A list of gene information.
}
\description{
NCBI Database API - Get single NCBI gene information
}
\details{
Get gene information form single NCBI gene ID.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
