% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{singleProteinInfo}
\alias{singleProteinInfo}
\title{Get single NCBI protein information}
\usage{
singleProteinInfo(proteinXml)
}
\arguments{
\item{geneXml}{Gene xml data.}
}
\value{
A list of protein information.
}
\description{
NCBI Database API - Get single NCBI protein information
}
\details{
Get gene information form single NCBI protein ID.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
