% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{singleTaxoInfo}
\alias{singleTaxoInfo}
\title{Get single NCBI taxonomy information}
\usage{
singleTaxoInfo(taxoXml)
}
\arguments{
\item{taxoXml}{Taxonomy xml data}
}
\value{
A matrix of taxonomy information
}
\description{
NCBI Database API - Get NCBI taxonomy information from given NCBI taxonomy IDs
}
\details{
Get taxonomy information form single NCBI taxonomy ID.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
