% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{getNCBITaxo}
\alias{getNCBITaxo}
\title{Get NCBI taxonomy information}
\usage{
getNCBITaxo(NCBITaxoIDs, n = 1, maxEach = 10000)
}
\arguments{
\item{NCBITaxoIDs}{A vector of NCBI taxonomy IDs.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{maxEach}{The maximum retrieve number in each visit. The ESearch, EFetch, and ESummary, the max number in one query is 10,000.}
}
\value{
A list containing taxonomy information for each ID.
}
\description{
NCBI Database API - Get NCBI taxonomy information from given NCBI taxonomy IDs
}
\details{
Get NCBI taxonomy information.
}
\examples{
## with two cores
tax3 <- getNCBITaxo(c('9606', '511145', '797302'), n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
Entrez Programming Utilities Help \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/}
}

