% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNCBI.R
\name{singleGenomeAnno}
\alias{singleGenomeAnno}
\title{Get single NCBI whole genomic gene annotation}
\usage{
singleGenomeAnno(genomeID, type = "CDS", n = 1)
}
\arguments{
\item{genomeID}{Single NCBI genome ID.}

\item{type}{"gene" or "CDS".}

\item{n}{The number of CPUs or processors, and the default value is 1.}
}
\value{
A list of annotation.
}
\description{
NCBI Database API - Get single NCBI whole genomic gene annotation
}
\details{
Get whole gene annotation form single NCBI genome ID. The locus tag is used as names for each gene.
}
\examples{
\dontrun{
aeuGenome <- singleGenomeAnno('CP007715', n = 4)}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

